Updated openstack/openstack Project: openstack-infra/storyboard  e42c6e38bf1acffeac732fd1980fd688b865b51a
One-To-One relationship for Access and Refresh Tokens

This changes the relationship in our access and refresh token table
from a many-to-many to a one-to-one. This is for two reasons:
Firstly, a many-to-many means it's possible to have a refresh token
issued that can refresh multiple access tokens, which actually
doesn't make a lot of sense, since our logic _also_ deletes all
access tokens associated with a used refresh token.

Secondly, many-to-many relationships do not support cascading updates
in SQLAlchemy, so our goal to manage all of our foreign relationships
in SQLA is not achievable if we maintain these kinds of relationships.

Change-Id: Iab3c65ac403961fccf7db15ef930f9d896626108