Updated openstack/openstack Project: openstack/horizon  b408104735e300db9d7d3e51c6f244209735c940
Adding styling for disabled and danger action items

We need styling for action-list menu-item for disabled and danger items.
Added attributes to li and a in menu-item for bootstrap to pick it up.
However, bootstrap does NOT support contextual styling of menu-items .
Basically, adding danger to the menu-item will not work. We have to add
this via our own css

Closes-bug: #1435627
Partially Implemenents: blueprint angularize-identity-tables
Change-Id: Ia38292e3968e94d17e2ae9b193bbbd28b02097ab