Updated openstack/openstack Project: openstack/cinder  17cc5cc53a3ecee743b65476210e65a66d0731c0
Verify all quotas before updating the database

Update quota multi-value in one request is not an atomic operation, so
even if a quota is invalid (e.g., volumes has a non-integer value),
some values may change but the response code is 400.

This patch will make sure the quota limits are of integer type for all
the requested quota keys before updating the database.

Closes-Bug: 1432972
Change-Id: If83af0a80489887098e52b03b9c4012e3ef061b5