Updated openstack/openstack Project: openstack/glance_store  47041aa6cc82692936c209bb1b5ada5ccdfb796e
Only warn on duplicate path on fs backend

Currently if the same path is used in two filesystem_store_datadirs a
BadStoreConfiguration exception is raised in the filesystem driver.

This is the desired behavior if the same path appears with different
priorities, as we don't know which one is correct, but if it appears
with the same priority (it is an exact duplicate) a warning in the log
should be enough.

DocImpact
Related-Bug: #1422699
Closes-Bug: #1428257
Change-Id: I4b0685f3c74851de6a47f1fdfdb6109b3eceb8ef