Updated openstack/openstack Project: openstack-infra/system-config  546578ea95d893b23b34c61c0a197174f53a0d80
Index cells logs

We've been trying to get the cells job in the check queue passing
consistently and it's been a game of whack-a-mole with regressions, so
when we're trying to debug new failures it'd be super helpful to
actually have logstash for the cells logs.

This adds the cells logs and only indexes them for the cells job(s)
which is the only time they should appear.

Change-Id: Id4450b7cb5d3303f9cb031c3e77fc17cfff97890