Updated openstack/openstack Project: openstack/horizon  71d441682fc9a9a1344b70f96ecbee99b3f6ccde
Removing unused variables in horizon angular

Before we can enable a global check for unused variables, we need to clean
up the code first. This way, the gate won't barf when we try to enable the
global unused and undef jshint config.

This patch addresses the files in angular horizon files.

Change-Id: I0495aa4d6a3dd16a3e6190e0187adf6069aa361a
Partially-Closes-bug: #1441794