Updated openstack/openstack Project: openstack/swift  583585086f6bbcbaada6ddf481424da9d24c3a95
Clarify the description of backward compatibility in Keystoneauth's docstring

If no prefix implies for the operator_roles and/or service_roles
parameters applies to all reseller_prefixes in the code.
But from backward comatibility point of view, the service_roles
parameter is newly introduced parameter so it is better to clarify
the docstring.

Change-Id: I8b0d5aeff785c21b8a8b0f5b20f774fad868016d