Updated openstack/openstack Project: openstack/nova  b9ca7c6a17af3931b290fbd68bca25b49b166eb1
Add min/max of API microversions to version API

As nova-spec api-microversions, versions API needs to expose minimum
and maximum microversions to version API, because clients need to
know available microversions through the API. That is very important
for the interoperability.
This patch adds these versions as the nova-spec mentioned.

Note:
  As v2(not v2.1) API change manner, we have added new extensions if
  changing API. However, this patch doesn't add a new extension even
  if adding new parameters "version" and "min_version" because version
  API is independent from both v2 and v2.1 APIs.

Closes-Bug: #1443375

Change-Id: Id464a07d624d0e228fe0aa66a04c8e51f292ba0c