Updated openstack/openstack Project: openstack/nova  e9a21ffda345bf3e95a6455d5cae5f2358e38a7f
VMware: Fix volume relocate during detach

During volume detach, volume's shadow VM might need to be
relocated if storage DRS moved the Nova instance along with
volume's vmdk. The shadow VM is relocated to the current
datastore of instance and it's virtual disk is updated. For
relocate, we are not setting the destination host and it
defaults to source host. This can result in relocate failure
if this host cannot access the destination datastore. This
patch sets the destination host correctly to that of instance.

Change-Id: If51bbef6e2ca811e2fffdca906645cc963c27a3e
Closes-Bug: #1425486