Updated openstack/openstack Project: openstack/tripleo-image-elements  c95c30df249c1a7047f651392fbadcf05c9ad389
load puppet overcloud heat templates and resource registry

This changes the element to use the new tuskar-load-role [1]
and tuskar-load-seed [2]. Includes the Compute and Controller
puppet roles as the extra-data files required by them. If the
puppet directory exists the puppet resource registry is used
with the puppet specific templates (tuskar roles).

[1] https://review.openstack.org/#/c/165150/
    (I9c143afb52c43e4258c3a6797a9707c08d8dcdaf)

[2] https://review.openstack.org/#/c/165285/
    (I07450edebd0dacd86bd0b28665eb1aaa3ea30ecb)

Change-Id: Id48d6531d71c09bc43df02f9df6cbbe2d21c3546