Updated openstack/openstack Project: openstack/tempest  faac59c79844765802b50330b7e9fe38f588b235
Deprecates 'format' in Javelin image description

Javelin uses the same field to describe disk format and container
format. This only works for ami images. This patch replaces the field
'format' in the resource file with 2 new fields: 'disk_format' and
'container_format'.

For backwards compatibility 'format' is still supported, but marked as
deprecated in the comments of Javelin.

Closes-bug: #1429909

Change-Id: Ic199956d11f8be974374b0f3c2cd7bdfedc90c8d