Updated openstack/openstack Project: openstack-infra/elastic-recheck  0f6ffe1bc062c483da92f8c868ce8785866ee19e
remove query for bug 1285323

After the modular grenade refactor we're generating a lot more
debugging information during service transitions. This means that this
bug was matching on most successful runs now.

The root bug is probably fixed now, lets assume so and will open
something new if it shows up.

Change-Id: I998a5cacab0e25fd7046e3cbba05c08ded012d3e
Closes-Bug: #1285323