Updated openstack/openstack Project: openstack-infra/project-config  a32f477f0d6e88d9cfdd8ec0a636e6a3174ed4ac
Configure stable branch ACLs for non-Oslo libs

Now that we are releasing stable branches of all of our libraries, we
need to configure the permissions in those branches to use the
stable-maint teams.

For projects with foo-stable-maint, use those teams.

For projects with foo-milestone, use those teams.

For projects with neither, amend the ACLs without making them exclusive
to add stable-maint-core and leave foo-core.

Also remove the blocks related to "proposed" branches, since we don't
use those for libraries.

Change-Id: I6fb79132d954c8cccdbd0c0584aa7acbe4655fef