Updated openstack/openstack Project: openstack/keystonemiddleware  c965851d86a75d17107bec82cf2796ab6a3724ae
Don't store expire into memcache

There are two times related to storing token values in memcache. The
memcache timeout and the expiry that is saved alongside the token. We
purposefully set the timeout to be longer than the expiry so that we can
cache that we have seen an expired token.

The expiry that is saved with the token is taken from within the token
body itself and tested for expiry before being returned from the cache.
This expiry is still present in the token body though and it means that
we have two different but exactly the same paths for where we test
expiry from.

Change the cache so that it always returns the full token body from the
cache. This way it is a drop in replacement for fetching the body from
keystone server or from within PKI and the token data is validated in
one place.

Change-Id: Ibe9816826e7d78b7ae25085f07ff9a2f18db9db0