Updated openstack/openstack Project: openstack/tooz  1c0fc10faa373cea43e4f8001d007b07d9472cfe
Allow the acquired file to be closed manually

Instead of requiring library users to wait for a GC
pass to close the file descriptor, make it possible
for a explicit close() call to do the same.

This makes it possible for users of these locks to
avoid running out of file descriptors that they are
unable to close.

Change-Id: I3fabc57b2f0f6b1cef45553cac7e04732a9c2489