Updated openstack/openstack Project: openstack-dev/grenade  9bf148268020ffe9e19c1048e75b5208c888f93d
ensure that grenade db is fresh on each run

The grenade db is really just an ini file, but there is no reason why
it should persist in a second run of grenade, and it actually causes
issues if it does. Blank it out in the init phase to ensure we're
starting with a blank store.

Change-Id: I22a2b53739153011f5c250739278719259ad7860