Updated openstack/openstack Project: openstack-infra/project-config  a13303442dec43d87a2f9e692c90e0a6cbad01af
remove stable-compat jobs from libs with stable branches

We don't need to run the backwards-compatibility tests for libraries
that have stable branches, because the master branch of the library will
never be used in the stable branch tests for applications.

Since the job macros are no longer used, remove them, too.

Change-Id: Ie38c5d213279c94a3aae5a69d2c30a3a67f0e391