Updated openstack/openstack Project: openstack/requirements  517f2174b025a4b7859e63308c3b58f4b4130e79
Disable syncs to libs to unjam master requirements changes

We need to disable syncs and checks of requirements in libraries so we
can release updated versions of all libs that have no capped
requirements. We can restore the syncs when that is done, and then we'll
look into ways to update the tools to not update caps on library
projects.

http://logs.openstack.org/24/173924/1/check/check-tempest-dsvm-full/6c09e06/logs/devstacklog.txt.gz#_2015-04-15_23_02_58_575

Change-Id: Ib948b756b8e6ca47a4c9c44c48031e54b7386a06