Updated openstack/openstack Project: openstack/neutron  8ca67832ea90b60d3b31d0a99159a0efa5f8d1ea
Increase max attempts to 2 for pings on ARP tests

If the server under test is under heavy load, the requirement
of the very first ping passing may be too strict. This patch
increases the max attempts to 2 for the ARP spoofing tests to
give time for the OVS flow changes to take effect.

Change-Id: Ib70790da23861a8ed9c77f9c11aaf8fa41bf581c
Closes-Bug: #1443916