Updated openstack/openstack Project: openstack/tempest  347defda0e9d7086392c8651054aa10ab1cc5fa5
FIx Accounts with identity v3 and v2 data in yaml

When tempest uses identity v3 - either because it's forced by a
test, or because it's configured as default, accounts can still
work with v2 accounts configured in accounts.yaml, under the
assumption that those accounts belong to
CONF.identity.admin_domain_name. The domain name is added on the
fly to newly created credentials by cred_provider, however this
is not accounted for in the get_hash function - fixing that.

Change-Id: Id0e72b9a8b85fdc682b9e14c39ecc1d8495ef35c