Updated openstack/openstack Project: openstack/nova  ca8faec7d1e360b9c1760bf42724d794bcfa5b10
Fix test_swap_volume_api_usage()

This test was triggering two attempts to revert the instance task state,
which was making a call to conductor that would never return because
there is no conductor running for this test. Since we were in a save-and-
reraise block, this just got ignored. Thus, this test used to take 120s
purely because it was waiting for two such attempts, at 60s each. Now it
takes (basically) zero time. Yay.

Change-Id: Ibe63cf4c47b3966dc95f70d5a0c9907ae5168264