Updated openstack/openstack Project: openstack/barbican  93e678077388e60b4faf7af5ec780ebd3eac2373
Add utility functions to convert between and PEM and DER

Adding utility routines to convert the format of public and private
keys between PEM and DER format, using only the pyOpenSSL and pyCrypto libraries.
These routines support PKCS#8.  The unit tests verify that the inputs and
outputs are identical with the equivalent openssl cli commands.

Change-Id: I27480088a5c441e1d2e66354cc08e9a27d7c3f18