Updated openstack/openstack Project: openstack/tempest  9536498aa7fbc4102ef436ef7ce1b0fa68ffbce8
Improve logging in fixed_networks

This commit makes some logging improvements to the common methods in
fixed_networks.py. Previously it was very difficult to trace through
any issues around network selection since it only would log the final
result and on failures. It also didn't attribute any messages to the
tests making the call. This fixes both issues by first leveraging
find_test_caller() from tempest-lib to get the test caller name and
secondly add more logging to help trace through the logic if something
fails. Additionally as part of this it was necessary to add additional
checks to the logic to ensure the logging was accurate and things were
in a known state.

Change-Id: Ifef67ed4578f48d7701635e7d81c144dc28b6120