Updated openstack/openstack Project: openstack-dev/devstack  44a969b110e452771cef48984b8e7f0964a44a34
Setting LOGFILE to a file in root directory wipes everything

clean.sh picks the parent of LOGFILE and wipes it clean! So if you
set it to a log file in the users root directory, you lose everything

We should delete just the LOGFILE and cleanup LOGDIR and SCREEN_LOGDIR if
they are explicitly set.

Change-Id: I45745427dcaed3dcf0b78cc9ed680833d9d555e8