Updated openstack/openstack Project: openstack/cinder  d7be849fcce5023eb2296ea322f490a1bf42ed18
SMBFS: Lock on a per-volume basis

This allows operations that do not conflict with each
other (i.e. are on different volumes) to run concurrently.
The prior locking scheme was too coarse and essentially
made the driver single-threaded.

This patch moves the implementation of the GlusterFS driver locking
scheme to the RemoteFS base driver so that other similar volume
drivers can use it.

Closes-Bug: #1439352
Change-Id: I37be14e7525406e0ad568f6228c1303998c5235f