Updated openstack/openstack Project: openstack/barbican  5c6b64c8c46b08d089d3b8013111dd7714ecf211
Potential resource exhaustion when registering consumers to containers

Previously, it was possible for consumers to be added with names of
arbitrary length. According to a finding by the Rackspace security
team, this ended up being a bigger problem than previously thought given
that addition of such consumers with sufficiently long names was apt to crash the running Barbican
instance.
Closes-Bug: 1430516

Change-Id: Ia5114d8de7dec796c56060892f389e47140c7834