Updated openstack/openstack Project: openstack/heat  e4447e7b66cf4691218b067643b5e25de72e31ba
Don't re-bind environment for get_file during stack update

While we're in the process of updating a stack, we set the stack's
environment to the new, updated values. However, we don't want to change
existing resources' idea of their own values until we have explivitly done
an update of them to bring them into line with the new
template/environment. So when parsing a template snippet containing the
get_file intrinsic, store the files dict to which it will refer. get_file
will resolve to the new file only when the template has been updated.

Closes-Bug: #1445170
Change-Id: I8ec32b0e38389cd314e4b2b2d8c65f61f8a9c41e