Updated openstack/openstack Project: openstack-infra/nodepool  3550c7c08f2afd669dc3f7634dcdfdd63e163d03
Use correct diskimage when building diskimages

When checking for missing DIB images if a missing image was found
nodepool would build that image. Problem was that it indexed the
diskimage config using the provider image's name not the provider
image's diskimage name. Update the key lookup to use the diskimage name.

Change-Id: I358c09a6d88ba34abcd0d42e37d40c5bb68e79c7