Updated openstack/openstack Project: openstack/barbican  faf315eb1516ad87e8fe385b3f4c1df62f284f08
Adding MySQL fixes to migrations

This change adds general fixes need for MySQL to perform the migrations
from stable/juno to kilo. For MySQL, the Alembic documentation
recommends adding the type to all drop_constraints calls, additionally
it is needed for alter_column as well. There is also an issue with MySQL
not allowing some alter_column changes on columns that have foreign key
constraints.

Changes
* correcting alter_column and drop_constraints in 795737bb3c3
* correcting alter_column in 256da65e0c5f
* correcting drop_constraints and associated behavior in 1c0f328bfce0
* adding names for the foreign constraints in ProjectSecret
* adding names for the foreign constraints in
  PreferredCertificateAuthorities

Change-Id: I558e5dca5d9c99150f2dceb251235e0d0a2f5a16
Closes-Bug: 1444209