Updated openstack/openstack Project: openstack/neutron  374be5d44b9d887533dbbcaadf5e8cb4d3bb0e34
Update decomp progress chart

This patch updates the progress chart, now that the first cycle after the
decomp started. For the fully decomposed plugins/drivers and for known
projects that integrate with  Neutron, this patch proposes a new
summary table that provides a go-to reference for everything Neutron related.

Related-blueprint: core-vendor-decomposition

Change-Id: Ib79a7b6d1401f1d9241621ae03cf6692685e12b1