Updated openstack/openstack Project: openstack-infra/project-config  e23348e59686dbfd7f69a85115ed168921989e06
Temporarily ignore requirements on master libs

We need to disable checks of master branch requirements in libraries
so we can release updated versions of all libs that have no capped
requirements. We can restore the checks when that is done, and then
we'll look into ways to update the tools to not update caps on
library projects.

http://logs.openstack.org/32/174532/1/check/gate-python-heatclient-requirements/ba4c6dd/console.html#_2015-04-16_19_03_49_535

Change-Id: I4e4ac0580b9a1a98cc5cc9ba62d6977324a15296