Updated openstack/openstack Project: openstack/neutron  26284228dfc3c5f121f869dd6b2d2a492afaf659
Get all interfaces for get_snat_sync_interfaces

The get_snat_sync_interfaces method was being called for
each router individually during a sync, which resulted
in a new query to the database.

This patch eliminates that waste by querying for the snat
interfaces for all of the routers in the list at once.

Change-Id: I1e44a0cf15a70632e8b62ac89ce807a7a457747d
Partial-Bug: #1445412