Updated openstack/openstack Project: openstack/oslo.config  7b1e157aeea426c58e3d4c9a76a231f0e5bc8241
Enforce alphabetic loading order for --config-dir contents

Deployers may want to rely on specific order of the files being loaded,
so we should make sure it's not changed, and that it's documented.

Note: test_config_dir already checks for the sorting order of
configuration files, so there is no need for a new test case.

Change-Id: Idc26b926601554b5592e13da92fbe81db99eb6e0