Updated openstack/openstack Project: openstack-dev/devstack  e350d3aa5312cbaae4b4595edc534867620a1dae
Assurance status of rabbitmq is running

I changed it so that rabbitmq always restart.

Current devstack don't restart rabbitmq in case of ubuntu. Because
rabbitmq is running at default.
But this approach have the following bug.
If rabbitmq is already installed and not running ,  stack.sh will
fail.
So I change it so that rabbitmq always restart.

Closes-bug: #1030798
Change-Id: Ie45446d3817b2f15631f03b2af84749fe936c67b