Updated openstack/openstack Project: openstack/nova  0d2fddbed09702f03b6c72dab8e396368884ce2e
Use list of requests in InstancePCIRequests.obj_from_db.

InstancePCIRequests.obj_from_db assumes it's called with with a dict
of values from instances_extra table, but in some cases it's called
with just the value of pci_requests column.
This changes obj_from_db to be used with just the value of pci_requests column.

Change-Id: I7bed733c845c365081719a70b8a2f0cc9a58370c
Closes-bug: #1445040