Updated openstack/openstack Project: openstack-dev/devstack  3894170067a4ca26952ef9b1315864c8dde8e1ad
Remove keystone public/admin_endpoint options

The public_endpoint and admin_endpoint options are used to set the base
hostname when listing versions. If they are not provided then the
response will use the same hostname as the request was made with - which
is almost always what you actually want.

This means devstack will respond correctly to a version list when behind
a floating IP for example.

Change-Id: Idc48b9d7bee9751deb24d730fdc560b163f39dfe