Updated openstack/openstack Project: openstack-infra/system-config  31f5e34f314a00dd745072731028947bbae2e912
Expose jenkins_gitfullname and jenkins_gitemail

Jenkins is consuming git user and email from jenkinsuser
parameters. But these parameters are not exposed on
OpenStack manifests.
Update all the manifests where it's relevant to send
that git username and email, either to jenkins slaves
or to static servers where jenkinsuser is needed.

Change-Id: I4e2b94b1220f88288401f9106721bc4df7fe9125