Updated openstack/openstack Project: openstack-infra/project-config  fd74b70abda808df89f786a788a04557912569e7
Fix retries in grab_console_log.sh

We implemented retries of console log grabbing to workaround occasional
jenkins 502 errors, but this is running under set -e so curl returning a
non zero error message resulted in short circuiting the retries and
failed the jobs.

Fix this with a simple || true appendage to the curl command so that any
failed curls can error properly without affecting the error checking in
the script. We will then retry until we eventually give up.

Change-Id: I0a784aa45bff7de5b411372ccecc569ec148ecc5