Updated openstack/openstack Project: openstack-dev/devstack  e5008e002b84afc3402877ea8b93729fa3a03401
Fix die_if_not_set tests

The "die_if_not_set" test has the LINENO as a positional argument.
The existing tests are not passing this in, so they are failing.

Along with this, remove the "-x" from the invocation and hide the
output of the tests that are expected to fail to avoid confusion.

Change-Id: Ibf6b9d7bb72b9f92831e1a90292ff8b0bec7faea