Updated openstack/openstack Project: openstack/tripleo-incubator  b5247cfd3e0ff4bdb32d0230909f60d7ca62bb3f
Don't hard code the use-ephemeral element

This makes it so that use-ephemeral isn't a hard coded requirement
rather it is just a default for DIB_COMMON_ELEMENTS.

This previous assertion (that use-ephemeral is required) is
no longer true with some TripleO configuration tools (like puppet).
When using Puppet use-ephemeral doesn't hurt anything but it
shouldn't be required... and could be confusing to some users.

Change-Id: I9bd11ce934deb5fd33b9a9c08f741aac216be8c8