Updated openstack/openstack Project: openstack/neutron  8793d3a40e5a67e410112ad18c13597ad32b5fb2
create_resource should return maximum length str

Previously, get_rand_name(max_length, prefix) returned a randomized
suffix integer which was concatenated to the end of the given prefix.
Effectively, the suffix was any decimal number between 1 and
0x7fffffff, so multiple calls to the function could return strings with
different length. This is unexpected since running an already
randomized name into the same function shouldn't return a different
string.

The suggested solution is to actually fill all the space needed until
the string is 'max_length' in size. Also, a check is added to
create_resource to make sure that it only generates a new port name if
the input prefix is less than the maximum device name and if the prefix
is long enough, don't generate a random port suffix.

Change-Id: I0d5a20c676f627bce2a377e3c451043150ca734c