Updated openstack/openstack Project: openstack/barbican  4635f4ef0a84d863c71ff08630a944ae177fbeed
Readability-related changes to secret store functions

This CR introduces minor changes with the aim to improve and clean the
code a little bit. So unnecessary comments were removed and some
functions renamed to reflect more context.

The need of the spec in the secret_store function in plugin.resources
was removed in favor of always getting a valid secret model.

On the other hand, there was an unnecessary check in the store_secret
function in the resources file. It was never the case (and there
shouldn't be a case) where the "spec" is null. So that check was
erased.

Change-Id: I73f59bdcc59ea213402f08862b87b48db8146ca0