Updated openstack/openstack Project: openstack/taskflow  6403d917f6a562a94a47f48f1575242d95a23914
Use the ability to chain exceptions correctly

In the zookeeper jobboard (and persistence backends) we
are associating the cause of a new exception when raising a
exception. Using the new exception helper we can make this
work better on py2.x and py3.x so that the py3.x version
has the chain setup correctly (while the py2.x version just
uses the class 'cause' attribute instead).

Change-Id: Ieeac2f70e1834d4612556565762ffd3be3e5b5a1