Updated openstack/openstack Project: openstack/python-openstackclient  2ea6f106918c6751c442a0515503838c03027ad2
Suppress warnings user can't fix

Requests/urllib3 started issuing warnings about certificates and SSL
that our users are unable to do anything about.  This is a very blunt
way to suppress these warnings unless --verbose or --debug is
supplied on the command line.

Being more precise in the suppression requires importing the warning
classes from urllib3 and dealing with the platforms where it has
been unvendored from requests.  Maybe in the future if there are
concerns that this mutes too much otherwise.

Change-Id: I50bb10a16222de12c5b95bfe042b92e43ea8ee7c